identifying FlowSteps by type
